[suiop] module refactor
